+++ ChangeLog.md
@@ -12,6 +12,7 @@ ChangeLog
 * #20: Allow callbacks to be used instead of `Element` classes in the `Reader`.
 * #25: Added `readText` to quickly grab all text from a node and advance the
   reader to the next node.
+* #15: Added `Sabre\XML\Element\Uri`.
 
 
 0.0.6 (2014-09-26)